+++ Kernel/Arch/SafeMem.h
