+++ Kernel/Arch/x86/common/Processor.cpp
@@ -20,13 +20,13 @@
 #include <Kernel/Thread.h>
 
 #include <Kernel/Arch/Processor.h>
+#include <Kernel/Arch/SafeMem.h>
 #include <Kernel/Arch/ScopedCritical.h>
 #include <Kernel/Arch/x86/CPUID.h>
 #include <Kernel/Arch/x86/InterruptDisabler.h>
 #include <Kernel/Arch/x86/Interrupts.h>
 #include <Kernel/Arch/x86/MSR.h>
 #include <Kernel/Arch/x86/ProcessorInfo.h>
-#include <Kernel/Arch/x86/SafeMem.h>
 #include <Kernel/Arch/x86/TrapFrame.h>
 
 #include <Kernel/Memory/PageDirectory.h>
