+++ Kernel/KSyms.cpp
@@ -5,8 +5,8 @@
  */
 
 #include <AK/TemporaryChange.h>
+#include <Kernel/Arch/SafeMem.h>
 #include <Kernel/Arch/SmapDisabler.h>
-#include <Kernel/Arch/x86/SafeMem.h>
 #include <Kernel/FileSystem/OpenFileDescription.h>
 #include <Kernel/KSyms.h>
 #include <Kernel/Process.h>
